% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_irda.R
\name{datatable_irda}
\alias{datatable_irda}
\title{Apply IRDA Theme to datatable Objects}
\usage{
datatable_irda(x, ...)
}
\value{
A datatable object.
}
\description{
This function applies a custom theme suitable for IRDA visualizations. It sets various graphical
parameters such as text size, font, and element colors.
}
\examples{
p <- datatable_irda(iris)

}
