% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarts_theme.R
\name{e_theme_irda}
\alias{e_theme_irda}
\title{Apply IRDA Theme to echarts Objects}
\usage{
e_theme_irda(x, ...)
}
\value{
A echarts object.
}
\description{
This function applies a custom theme suitable for IRDA visualizations. It sets various graphical
parameters such as text size, font, and element colors.
}
\examples{
iris|>
dplyr::group_by(Species) |>
  echarts4r::e_charts(Sepal.Length) |>
  echarts4r::e_scatter(Sepal.Width) |>
  e_theme_irda()
}
