% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_irda.R
\name{flextable_irda}
\alias{flextable_irda}
\title{Apply IRDA Theme to flextable Objects}
\usage{
flextable_irda(x, ...)
}
\value{
A flextable object.
}
\description{
This function applies a custom theme suitable for IRDA visualizations. It sets various graphical
parameters such as text size, font, and element colors.
}
\examples{
p <- flextable::flextable(iris) |>
  flextable_irda()

}
