% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irda_palette.R
\name{irda_palette}
\alias{irda_palette}
\title{Retrieve IRDA Color Palettes}
\usage{
irda_palette(palette = "main")
}
\arguments{
\item{palette}{Name of the palette to retrieve, defaults to 'main'.}
}
\value{
A vector of color hex codes.
}
\description{
This function provides a selection of color palettes designed for IRDA visualizations.
}
\examples{
colors <- irda_palette("main")
}
