% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_gen.R
\name{palette_gen}
\alias{palette_gen}
\title{Generate a Color Palette Function}
\usage{
palette_gen(palette = "main", direction = 1)
}
\arguments{
\item{palette}{Name of the palette to use, defaults to 'main'.}

\item{direction}{Direction to arrange colors, 1 for normal, -1 for reversed.}
}
\value{
A function that generates color palettes.
}
\description{
This function generates a function that returns a specified number of colors from a given palette,
potentially reversing the direction of the color sequence.
}
\examples{
color_function <- palette_gen("main")
color_function(3)
}
