% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_irda.R
\name{scale_color_irda}
\alias{scale_color_irda}
\title{Scale Color with IRDA Palettes}
\usage{
scale_color_irda(palette = "main", direction = 1, ...)
}
\arguments{
\item{palette}{Name of the color palette to use, defaults to 'main'.}

\item{direction}{Direction to arrange colors, 1 for normal, -1 for reversed.}

\item{...}{Other parameters passed to discrete_scale.}
}
\value{
A ggplot2 scale function.
}
\description{
Provides a ggplot2 scale function for color aesthetics using IRDA color palettes.
}
\examples{
ggplot(data = iris, aes(x = Sepal.Length, color = Species)) +
  geom_point() +
  scale_color_irda()
}
