% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_irda.R
\name{scale_fill_irda}
\alias{scale_fill_irda}
\title{Scale Fill Color with IRDA Palettes}
\usage{
scale_fill_irda(palette = "main", direction = 1, ...)
}
\arguments{
\item{palette}{Name of the color palette to use, defaults to 'main'.}

\item{direction}{Direction to arrange colors, 1 for normal, -1 for reversed.}

\item{...}{Other parameters passed to discrete_scale.}
}
\value{
A ggplot2 scale function.
}
\description{
Provides a ggplot2 scale function for fill aesthetics using IRDA color palettes.
}
\examples{
ggplot(data = iris, aes(x = Sepal.Length, fill = Species)) +
  geom_histogram(stat = "count", position = "dodge") +
  scale_fill_irda()
}
