% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_irda.R
\name{theme_gt_irda}
\alias{theme_gt_irda}
\title{Apply IRDA Theme to gt Objects}
\usage{
theme_gt_irda(x, ...)
}
\value{
A gt object.
}
\description{
This function applies a custom theme suitable for IRDA visualizations. It sets various graphical
parameters such as text size, font, and element colors.
}
\examples{
p <- gt::gt(iris) |>
     theme_gt_irda()
}
